Bugfix: user can tap on the blank space on the show answer area in QuestionGame.
- Refactoring of QuestionGameFragment tests. I thought that the tests violated the DRY principle, and I ended up having to change stuff in a lot of places because I changed some text on the string files.
This smelled, so I created a QuestionGameScreen class that encapsulated all the manipulation of the screen, and both questionGame test classes could then use this encapsulated class.
Now it looks like the code is much more maintainable, and I can extend the concept and reuse some parts of the QuestionGameScreen class (by extracting the common behaviour to a superclass, perhaps) in other tests should the need arise.
Eventually, I will do the same thing for acceptance testing.

-Fixed the problem that it was not possible to click on the blank parts on the screen in click Here to show answer. The way to fix that was to inflate parts of the screen to the container as needed. So when I need to show the click here to show answer I inflate a TextView but when the user clicks on the box a ScrollView is inflated.
Also, the clickListener was set on the frameLayout (the container), that expands to the whole available space.