Fixed automation tests on Jenkins.
Automation tests broke and I didn't get them on my machine because I was using xoom.
I refactored part of the automation tests to use the same strategy that I used on unit tests. In order to do that,
now the QuestionGameScreen receives a automator, some class whose responsibility is to execute actions.
Depending on which type of test is running, this automator can either use the default method calls (in the case of unit test) or use solo (in acceptance tests) to do what it needs to do. In any case, only the logic of checking stuff stays in the screen (it should be the same for both tests) and methods that delegates to the automators.