Usability: Add new Question line on List Questions TableView.
Now there is a Add new Question line on the tableView that lists questions for android versions prior to Honeycomb.
This was made because it was too boring to have to click in the menu and click on add every time someone wanted to create a new question. On Versions prior to honeycomb the application does not show an action bar.