Refactoring: Tests of QuestionEditFragment.
QuestionEdit*FragmentTest are now depending on QuestionEditScreen. Instead of QuestionEditScreen depending on Automators, now there's a class called TestIds that can be used to give ids for acceptance or unit testing, so there is no need for screens to declare their own automators anymore.
The only screen that still has its own automators is questionGameScreen, but this structure is deprecated and will vanish somewhere in the future.