Benchmarking: QuestionGameActivity
I Benchmarked the QuestionGame to see how much time it took to rebuild the queue, and if the user could use the app even if he put 1000 questions in the database. Turns out for now it works fine.

Also refactored the QuestionListFragmentTest to use QuestionListScreen, similar to what I've done previously with QuestionEditFragment and QuestionGameFragment.