Refactoring: Automator => Mechanize
I moved a bunch of stuff from its place and did some refactorings on the screens.
Ultimately, what I wanted to do (move stuff to intern and run acceptance tests on android 4.x) did not work.
Also, intern-android is now an apklib instead of an apk.

Note: Seems like maven-android-plugin doesn't like transitive dependencies too well. Tried to do both using the apklib and app/jar dependency structure, none worked. Either it fails to find the dependencies of the now non-existent android-testlib (which contained the screens base classes) or it failed to invoke methods on intern-android-test because of jam/dalvikvm lack of compatibility. Anyway, I could not fix it, so I did some YAGNI and the classes that would go to android-testlib stayed in Kotoba.